adds incident serializers service object
